torque: modify value used for resources_max.nodect and resources_available.nodect
resources_available.nodect = N torque config parameter, that is used by
ParallelCluster to set the max size that the cluster can assume, is not
behaving exactly as documented, but it's limiting the number of jobs
that can run concurrently by enforcing the total number of nodes to N
even if a node has free slots to be allocated.

resources_available.nodect = N serves two different causes:
1. Enables torque to accept jobs that request a number of nodes greater
than the ones available to the cluster (nodes that the scaling daemons
will launch)
2. Prevents the submission of jobs that exceed the max size of the cluster
The workaround for the following issue consists in setting
resources_max.nodect = max_queue_size, so that no jobs are accepted if
max_queue_size is exceeded, and change
resources_available.nodect = max_queue_size*node_slots so that torque will
correctly use all available slots in the cluster.

Note that sqswatcher will take care of setting the correct value for
resources_available.nodect.

Signed-off-by: Francesco De Martino <fdm@amazon.com>