Expose all-or-nothing option in Slurm resume config
When set to True, instance provisioning will follow an all-or-nothing strategy
within the configured max_batch_size. It means that ParallelCluster will try to
launch all requested nodes at once (all-or-nothing) in batches of max_batch_size.
This all or nothing scaling is not job based but is applied to all instances requested
by Slurm when invoking the resume program. Also the all or nothing provisioning is
independent for each compute resource.

Signed-off-by: Francesco De Martino <fdm@amazon.com>